feat: Add `addr` feature
`interface_and_mtu` is now behind the default-on `addr` feature.

The point of this is to introduce `interface_and_mtu_of_socket`, which
operates on an existing `UdpSocket`. That means we can avoid `connect`,
which is blocklisted by Firefox for dependencies.

The downside is that we would need to pass in the actual `UdpSocket` to
the neqo transport code, which is probably a huge pain...